refactor: address feedback from @t1m0thyj
