+++ packages/zowe-explorer/__tests__/__unit__/extension.unit.test.ts
@@ -191,7 +191,6 @@ async function createGlobalMocks() {
             "zowe.uss.refreshDirectory",
             "zowe.uss.fullPath",
             "zowe.uss.editSession",
-            "zowe.uss.ZoweUSSNode.open",
             "zowe.uss.removeSession",
             "zowe.uss.createFile",
             "zowe.uss.createFolder",
