--- packages/zowe-explorer/i18n/sample/src/dataset/DatasetFSProvider.i18n.json
@@ -1,3 +1,0 @@
-{
-  "localize.uss.profileNotFound": "Profile does not exist for this file."
-}
