--- packages/zowe-explorer/i18n/sample/src/job/init.i18n.json
@@ -1,3 +1,0 @@
-{
-  "jobActions.fetchSpoolFile": "$(sync~spin) Fetching spool files..."
-}
