+++ packages/zowe-explorer/package.json
@@ -2104,6 +2104,7 @@
     "copy-webpack-plugin": "^12.0.2",
     "cross-env": "^5.2.0",
     "del": "^4.1.1",
+    "disposablestack": "^1.1.4",
     "eslint-plugin-zowe-explorer": "3.0.0-next-SNAPSHOT",
     "event-stream": "^4.0.1",
     "expect": "^24.8.0",
@@ -2123,7 +2124,6 @@
     "@zowe/zosmf-for-zowe-sdk": "8.0.0-next.202403132009",
     "@zowe/zowe-explorer-api": "3.0.0-next-SNAPSHOT",
     "dayjs": "^1.11.10",
-    "disposablestack": "^1.1.3",
     "fs-extra": "8.0.1",
     "isbinaryfile": "4.0.4",
     "promise-queue": "2.2.5"
