+++ packages/zowe-explorer/src/SpoolUtils.ts
