+++ packages/zowe-explorer/src/globals.ts
@@ -23,7 +23,7 @@ export let USS_DIR: string;
 export let DS_DIR: string;
 export let CONFIG_PATH; // set during activate
 export let LOG: imperative.Logger;
-export const COMMAND_COUNT = 125;
+export const COMMAND_COUNT = 124;
 export const MAX_SEARCH_HISTORY = 5;
 export const MAX_FILE_HISTORY = 10;
 export const MS_PER_SEC = 1000;
